app loads new bluetooth fragment from both the menu and through clicking the device icon on the home page.
I will be working on implementing search, and connect functionality for ble devices within the bluetooth manager fragment. the example I am referencing begins a new activity once connection is selected. I am unsure as to how to handle that but it "should" display a confirmation of connection on bluetooth fragment and users should be able to continue through the app (maybe bluetooth connection and maintenance of connection is handle by a thread in the main activity??)